Update source code script and name space
Previously we had problems with name space and masking.

I went through all the non-generic functions that we used in the script
and added package names in front of them. This resolved the names space
issue. The list of the functions are recorded in a word document. Let
me know if you would like to include it on the GitHub page.

For masking, I rewrote the code, so that we don’t have to use “fixef”
and “VarCorr” from multiple packages. Only cplm::VarCorr remains, since
there is clear way to extract the information from regression fit
directly. As for the package imports, we now only import the functions
we use. This gets rid of masking issue.

There was a typo in GetBootCI that uses cplm.data. It’s now been
removed. Since we do not need this variable, I also removed it from
globalVariables (the if statement in the beginning of the R script).

I also added the FitCPMM and GetBootCI examples back.